requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
